Move existing channel state/reducer to app-channel
