+++ packages/wallet/src/containers/Closing.tsx
@@ -3,7 +3,7 @@ import { PureComponent } from 'react';
 import { connect } from 'react-redux';
 import { Button } from 'reactstrap';
 
-import * as states from '../redux/channelState/state';
+import * as states from '../redux/channelState/app-channel/state';
 import * as actions from '../redux/actions';
 
 import AcknowledgeX from '../components/AcknowledgeX';
