+++ packages/wallet/src/redux/__tests__/test-scenarios.ts
@@ -1,7 +1,8 @@
 import { Channel, CommitmentType, Commitment } from 'fmg-core';
 import { channelID } from 'fmg-core/lib/channel';
 import { bigNumberify } from 'ethers/utils';
-import { waitForPreFundSetup } from '../channelState/state';
+import { waitForPreFundSetup } from '../channelState/app-channel/state';
+import { ChannelType } from '../channelState/shared/state';
 
 export const libraryAddress = '0x' + '1'.repeat(40);
 export const channelNonce = 4;
@@ -119,5 +120,9 @@ export const initializedChannelState = {
   }),
 };
 export const initializingChannelState = {
-  [asAddress]: { address: asAddress, privateKey: asPrivateKey },
+  [asAddress]: {
+    address: asAddress,
+    privateKey: asPrivateKey,
+    channelType: 'Application' as ChannelType,
+  },
 };
