+++ packages/wallet/src/redux/channelState/app-channel/state.ts
@@ -1,0 +1,25 @@
+import { OpeningState } from './opening/state';
+import { RunningState } from './running/state';
+import { FundingState } from './funding/state';
+import { ChallengingState } from './challenging/state';
+import { RespondingState } from './responding/state';
+import { WithdrawingState } from './withdrawing/state';
+import { ClosingState } from './closing/state';
+
+export type OpenedState =
+  | FundingState
+  | RunningState
+  | ChallengingState
+  | RespondingState
+  | WithdrawingState
+  | ClosingState;
+
+export type AppChannelStatus = OpeningState | OpenedState;
+
+export * from './opening/state';
+export * from './running/state';
+export * from './funding/state';
+export * from './challenging/state';
+export * from './responding/state';
+export * from './withdrawing/state';
+export * from './closing/state';
