+++ packages/wallet/src/redux/channelState/state.ts
@@ -1,43 +1,22 @@
-import { OpeningState } from './opening/state';
-import { RunningState } from './running/state';
-import { FundingState } from './funding/state';
-import { ChallengingState } from './challenging/state';
-import { RespondingState } from './responding/state';
-import { WithdrawingState } from './withdrawing/state';
-import { ClosingState } from './closing/state';
+import { AppChannelStatus } from './app-channel/state';
+import { ChannelType } from './shared/state';
 
 export interface InitializingChannelStatus {
   address: string;
   privateKey: string;
+  channelType: ChannelType;
 }
 
 export interface InitializingChannelState {
   [participantAddress: string]: InitializingChannelStatus;
 }
 
-export type OpenedState =
-  | FundingState
-  | RunningState
-  | ChallengingState
-  | RespondingState
-  | WithdrawingState
-  | ClosingState;
-
-export type ChannelStatus = OpeningState | OpenedState;
-
+export type ChannelStatus = AppChannelStatus;
 export interface InitializedChannelState {
-  [channelId: string]: ChannelStatus;
+  [channelId: string]: AppChannelStatus;
 }
 export interface ChannelState {
   initializingChannels: InitializingChannelState;
   initializedChannels: InitializedChannelState;
   activeAppChannelId?: string;
 }
-
-export * from './opening/state';
-export * from './running/state';
-export * from './funding/state';
-export * from './challenging/state';
-export * from './responding/state';
-export * from './withdrawing/state';
-export * from './closing/state';
